+++ plotter/src/main/java/com/devsmart/plotter/Function.java
@@ -2,5 +2,5 @@ package com.devsmart.plotter;
 
 public interface Function {
 
-    double value(double x);
+    float value(float x);
 }
